requirements.txt
# ============================================================
#  Multi-Agent Assistant — Python Dependencies
# ============================================================

# Core LangChain framework
langchain>=0.2.0
langchain-core>=0.2.0
langchain-google-genai>=1.0.0

# Google Gemini LLM
google-generativeai>=0.5.0

# Streamlit frontend
streamlit>=1.35.0

# HTTP requests (Weather API)
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0
